fix: forge fmt

+++ test/util/ZkEVMCommon.sol
@@ -6,22 +6,13 @@ import "forge-std/Test.sol";
 abstract contract ZkEVMCommon is Test {
     string constant MERKLE_TREE_HEIGHT = "32"; // As of now, the height of the Merkle tree is fixed to 32
 
-    function _encodeLeaves(
-        bytes32[] memory leaves
-    ) public pure returns (string memory encodedLeaves) {
-        for (uint i = 0; i < leaves.length; i++) {
-            encodedLeaves = string(
-                abi.encodePacked(
-                    encodedLeaves,
-                    _bytes32ToHex(abi.encodePacked(leaves[i]))
-                )
-            );
+    function _encodeLeaves(bytes32[] memory leaves) public pure returns (string memory encodedLeaves) {
+        for (uint256 i = 0; i < leaves.length; i++) {
+            encodedLeaves = string(abi.encodePacked(encodedLeaves, _bytes32ToHex(abi.encodePacked(leaves[i]))));
         }
     }
 
-    function _getMerkleTreeRoot(
-        string memory encodedLeaves
-    ) public returns (bytes32) {
+    function _getMerkleTreeRoot(string memory encodedLeaves) public returns (bytes32) {
         string[] memory operation = new string[](5);
         operation[0] = "node";
         operation[1] = "tools/zkevm-commonjs-wrapper";
@@ -33,10 +24,7 @@ abstract contract ZkEVMCommon is Test {
         return abi.decode(result, (bytes32));
     }
 
-    function _getProofByIndex(
-        string memory encodedLeaves,
-        string memory index
-    ) public returns (bytes32[32] memory) {
+    function _getProofByIndex(string memory encodedLeaves, string memory index) public returns (bytes32[32] memory) {
         string[] memory operation = new string[](6);
         operation[0] = "node";
         operation[1] = "tools/zkevm-commonjs-wrapper";
@@ -49,9 +37,7 @@ abstract contract ZkEVMCommon is Test {
         return abi.decode(result, (bytes32[32]));
     }
 
-    function _bytes32ToHex(
-        bytes memory buffer
-    ) internal pure returns (string memory) {
+    function _bytes32ToHex(bytes memory buffer) internal pure returns (string memory) {
         bytes memory converted = new bytes(buffer.length * 2);
 
         bytes memory _base = "0123456789abcdef";
